XMLパース機能を関数からクラスに変更
ViewModelのテストがやや書きづらいため
